jadx-0.8.0 - B2 v5.2.0 (1467)

+++ sources/org/telegram/messenger/voip/VoIPController.java
@@ -11,9 +11,10 @@ import java.util.Calendar;
 import java.util.Iterator;
 import java.util.Locale;
 import org.telegram.messenger.ApplicationLoader;
+import org.telegram.messenger.BuildVars;
 import org.telegram.messenger.MessagesController;
-import org.telegram.p005ui.Components.voip.VoIPHelper;
 import org.telegram.tgnet.TLRPC.TL_phoneConnection;
+import org.telegram.ui.Components.voip.VoIPHelper;
 
 public class VoIPController {
     public static final int DATA_SAVING_ALWAYS = 2;
@@ -145,7 +146,7 @@ public class VoIPController {
         int a = 0;
         while (a < endpoints.length) {
             TL_phoneConnection endpoint = endpoints[a];
-            if (endpoint.var_ip == null || endpoint.var_ip.length() == 0) {
+            if (endpoint.ip == null || endpoint.ip.length() == 0) {
                 throw new IllegalArgumentException("endpoint " + endpoint + " has empty/null ipv4");
             } else if (endpoint.peer_tag == null || endpoint.peer_tag.length == 16) {
                 a++;
@@ -238,6 +239,8 @@ public class VoIPController {
     }
 
     public void setConfig(double recvTimeout, double initTimeout, int dataSavingOption, long callID) {
+        String logFilePath;
+        String logFilePath2;
         ensureNativeInstance();
         boolean sysAecAvailable = false;
         boolean sysNsAvailable = false;
@@ -252,7 +255,17 @@ public class VoIPController {
         long j = this.nativeInst;
         boolean z = (sysAecAvailable && VoIPServerConfig.getBoolean("use_system_aec", true)) ? false : true;
         boolean z2 = (sysNsAvailable && VoIPServerConfig.getBoolean("use_system_ns", true)) ? false : true;
-        nativeSetConfig(j, recvTimeout, initTimeout, dataSavingOption, z, z2, true, getLogFilePath(callID), null, false);
+        if (BuildVars.DEBUG_VERSION) {
+            logFilePath = getLogFilePath("voip" + callID);
+        } else {
+            logFilePath = getLogFilePath(callID);
+        }
+        if (BuildVars.DEBUG_VERSION && dump) {
+            logFilePath2 = getLogFilePath("voipStats");
+        } else {
+            logFilePath2 = null;
+        }
+        nativeSetConfig(j, recvTimeout, initTimeout, dataSavingOption, z, z2, true, logFilePath, logFilePath2, BuildVars.DEBUG_VERSION);
     }
 
     public void debugCtl(int request, int param) {
@@ -289,20 +302,22 @@ public class VoIPController {
 
     private String getLogFilePath(long callID) {
         File dir = VoIPHelper.getLogsDir();
-        File[] _logs = dir.listFiles();
-        ArrayList<File> logs = new ArrayList();
-        logs.addAll(Arrays.asList(_logs));
-        while (logs.size() > 20) {
-            File oldest = (File) logs.get(0);
-            Iterator it = logs.iterator();
-            while (it.hasNext()) {
-                File file = (File) it.next();
-                if (file.getName().endsWith(".log") && file.lastModified() < oldest.lastModified()) {
-                    oldest = file;
+        if (!BuildVars.DEBUG_VERSION) {
+            File[] _logs = dir.listFiles();
+            ArrayList<File> logs = new ArrayList();
+            logs.addAll(Arrays.asList(_logs));
+            while (logs.size() > 20) {
+                File oldest = (File) logs.get(0);
+                Iterator it = logs.iterator();
+                while (it.hasNext()) {
+                    File file = (File) it.next();
+                    if (file.getName().endsWith(".log") && file.lastModified() < oldest.lastModified()) {
+                        oldest = file;
+                    }
                 }
+                oldest.delete();
+                logs.remove(oldest);
             }
-            oldest.delete();
-            logs.remove(oldest);
         }
         return new File(dir, callID + ".log").getAbsolutePath();
     }
